refactor: change Bignum.length data type from int to unsigned long long int
This is so that we can have a bigger range for bignum, as Bignum.length tracks the number of digits of the Bignum, and unsigned long long int has a wider range. This also ensures that Bignum.length can only be a positive integer

+++ bignum.h
@@ -23,7 +23,8 @@ typedef struct {
     // Integer will be in reverse for easier arithmetic
     int digits[MAX_INT_LENGTH];
     // Number of digits in integer
-    int length;
+    // Bignum.length will cap out at around 18,446,744,073,709,551,615 (ref: C docs). Therefore the Bignum can only have a maximum of 18,446,744,073,709,551,615 digits.
+    unsigned long long int length;
     // Sign will be 0 (positive) or 1 (negative), following the binary sign bit convention.
     BIGNUM_SIGN sign;
 } Bignum;
